For you can see the files:
First, you need to install the app where the files are, the "Servidor" called XAMPP, and next to that install WordPress.org, the version free. Installed that, open XAMPP and "run" Apache and MySQL clicked in "start". Then you open the navegator and put "localhost/gietc/admin", put your username GIETC and the password that you saved in notion, and thats it, you can make changes on the page.